Add some lines for building linux-32 conda binaries
There is currently a problem with linux-32 anaconda gcc-4.8.5 package
So I have changed it to 4.8.2 for linux-32 builds.  Also there is a library
linking problem specific to using gcc in a conda-build, but not with installing
gcc.  So we have to copy the library files over for the build to work...
To do so just comment or uncomment the lines in build.sh